Gigantic commit implementing several bugfixes and features.
This is generally verboten, but the two main features feed into
eachother, and the bugfixes are borne out of that.

First, it implements the 'caller args' feature -- delayed expressions
now see the @_ that was in scope when they were delayed.

Second, it implements thread safety, for both the module and delayed
arguments.

The most notable bugfixes include:

* Delaying the creation of coderefs that close over lexicals now work:
    my $foo; sub { delay $foo .= 10 }->()

* A delayed eval STRING no longer sees the delayer's lexicals

* A delayed __SUB__ works properly